amd_cloud_files/requirements.txt
# AMD Cloud Pipeline Dependencies
# PyTorch with ROCm should already be installed:
#   pip install torch torchvision --index-url https://download.pytorch.org/whl/rocm6.1

fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
requests>=2.31
Pillow>=10.0
numpy>=1.24
trimesh>=4.0
scipy>=1.11
open3d>=0.18.0

# These are needed by the model repos but should already be installed:
# transformers
# segment-anything
# mmengine
# mmsegmentation
